// src/common/constants/feature.constant.d.ts
export declare enum FeatureKey {
    MCP_SERVER = "mcpServer",
    GIT_SYNC = "gitSync",
    SCHEMA_REPORT = "schemaReport",
    DRIFT_DETECTION = "driftDetection",
    BETA_GATE = "betaGate",
    BUG_REPORT = "bugReport",
    FEATURE_SUGGEST = "featureSuggest",
    DB_DESIGNER = "dbDesigner",
    AUTO_BACKUP = "autoBackup",
    FOCUS_COLUMN_MODE = "focusColumnMode",
    ER_DIAGRAM = "erDiagram",
    QUICK_DUMP_COMPARE = "quickDumpCompare",
    LIVE_DEMO = "liveDemo",
    DOMAIN_LOGIC = "domainLogic",
    RESTRICTED_USER = "restrictedUser",
    SNAPSHOT_RESTORE = "snapshotRestore"
}